requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.6.0
beautifulsoup4==4.11.2
branca==0.6.0
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.2
decorator==5.1.1
distlib==0.3.6
dj-database-url==0.5.0
dj-email-url==1.0.6
Django==4.2
django-cache-url==3.4.4
django-phonenumber-field==6.0.0
django_debug_toolbar==3.8.1
environs==5.2.1
filelock==3.9.0
flake8==6.0.0
folium==0.14.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
isoweek==1.3.3
Jinja2==3.1.2
livereload==2.6.3
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
more-itertools==9.1.0
numpy==1.24.2
packaging==23.0
pathlib==1.0.1
python-dotenv==1.0.0
python-telegram-bot==13.15
requests==2.28.2

